% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_sites.R
\name{binding_sites}
\alias{binding_sites}
\title{Get the binding sites of CLIP-seq.}
\usage{
binding_sites(assembly = "hg19", datasetID)
}
\arguments{
\item{assembly}{=[genome version]. hg19, mm10, dm6, ce10, sacCer3}

\item{datasetID}{=[uniq dataset ID]. e.g., SBDH27}
}
\value{
chrom - The name of the chromosome (e.g. chr3, chrY, chr2_random) or scaffold (e.g. scaffold10671).

chromStart - The starting position of the feature in the chromosome or scaffold. The first base in a chromosome is numbered 0.

chromEnd - The ending position of the feature in the chromosome or scaffold.

name - Defines the name of the BED line. This label is displayed to the left of the BED line in the Genome Browser window when the track is open to full display mode or directly to the left of the item in pack mode.

score - A score between 0 and 1000. If the track line useScore attribute is set to 1 for this annotation data set, the score value will determine the level of gray in which this feature is displayed (higher numbers = darker gray).

strand - Defines the strand. Either "." (=no strand) or "+" or "-".
}
\description{
Download the binding sites of CLIP-seq in bed format
}
\examples{

bind <- binding_sites(datasetID = "SBDH27")

}
