% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNA_fun.R
\name{ceRNA_fun}
\alias{ceRNA_fun}
\title{Get Data of the CeRNA Networks}
\usage{
ceRNA_fun(
  assembly = "hg19",
  geneType,
  ceRNA = "all",
  miRNAnum = 2,
  family,
  pval = 0.01,
  fdr = 0.01,
  pancancerNum = 0
)
}
\arguments{
\item{assembly}{=[genome version]: hg19, mm10}

\item{geneType}{=[main gene type]: mRNA, lncRNA, pseudogene}

\item{ceRNA}{=[gene name]. e.g., MYC ("all" for downloading all regulatory data)}

\item{miRNAnum}{=[integer]: miminum number of miRNAs}

\item{family}{=[miRNA family name]. e.g., miR-200bc-3p/429,miR-513c-5p/514b-5p("all" for downloading all regulatory data)}

\item{pval}{=[decimal]: maximum p-value (<=0.01).}

\item{fdr}{=[decimal]: maximum false dicovery rate (<=0.01)}

\item{pancancerNum}{=[integer]: minimum number of Cancer types (Pan-Cancer, <=32)}
}
\value{
geneID	ENSEMBL gene ID, e.g., ENSG00000148584

geneName	Name of gene, e.g., A1CF

geneType	Type of gene, e.g., protein_coding

ceRNAid	ENSEMBL gene ID, e.g., ENSG00000018510

ceRNAname	Name of gene, e.g., AGPS

ceRNAgeneType	Type of gene, e.g., protein_coding

hitMiRNAFamilyNum	Number of hit miRNA families

hitMiRNAFamily	Name of hit miRNA family

pval	P value

fdr	False dicovery rate

pancancerNum	Number of Cancer types (r<0, p-value<0.05)
}
\description{
Retrieve the ceRNA networks from thousands of interactions of miRNA-targets supported by CLIP-seq data.
}
\examples{
ceRNA <- ceRNA_fun(geneType = "mRNA", family = "miR-200bc-3p/429,miR-513c-5p/514b-5p")

}
