% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Degradome-RNA.R
\name{degradome_RNA}
\alias{degradome_RNA}
\title{Get Data for MiRNAs Cleavage Events}
\usage{
degradome_RNA(
  assembly = "hg19",
  geneType,
  miRNA = "all",
  ampdegraExpNum = 1,
  target,
  cellType = "all"
)
}
\arguments{
\item{assembly}{=[genome version]. e.g., hg19}

\item{geneType}{=[main gene type]: mRNA, ncRNA}

\item{miRNA}{=[microRNA name]. e.g., hsa-miR-196a-5p ("all" for downloading all regulatory data)}

\item{ampdegraExpNum}{=[integer]: minimum number of supporting degradome-seq experiments}

\item{target}{=[gene name]. e.g., TP53 ("all" for downloading all regulatory data)}

\item{cellType}{=[cell type]. e.g., HeLa ("all" for downloading all regulatory data)}
}
\value{
A data frame object

miRNAid	MiRBase microRNA ID, e.g., MIMAT0002175

miRNAname	Name of microRNA, e.g., hsa-miR-485-5p

geneName	Name of gene, e.g., TP53

geneType	Type of gene, e.g., protein_coding

cleaveEventNum	Number of cleavage events

degraExpNum	Number of supporting degradome-seq experiments

degraSiteNum	Number of cleavage sites from supporting degradome-seq experiments

totalReads	Number of total reads from all samples

category	Defined by starscan program
}
\description{
Retrieve data for cleavage events of miRNAs on genes supported by degradome-seq data
}
\examples{

dd_tp_brca <- degradome_RNA(geneType = "mRNA", target = c("TP53", "BRCA1"))
dd_h19 <- degradome_RNA(target = "H19", geneType = "ncRNA")
}
