% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_target.R
\name{mirna_target}
\alias{mirna_target}
\title{Get MiRNA Target Data}
\usage{
mirna_target(
  assembly = "hg19",
  geneType,
  miRNA = "all",
  clipExpNum = NULL,
  ampdegraExpNum = NULL,
  pancancerNum = NULL,
  programNum = NULL,
  program = NULL,
  target,
  cellType = "all"
)
}
\arguments{
\item{assembly}{genome version: hg19, mm10}

\item{geneType}{main gene type: mRNA, lncRNA, pseudogene, circRNA, sncRNA}

\item{miRNA}{microRNA name. e.g., hsa-miR-21-5p; "all" for downloading all regulatory data}

\item{clipExpNum}{integer: minimum number of supporting CLIP-seq experiments}

\item{ampdegraExpNum}{integer: minimum number of supporting degradome-seq experiments}

\item{pancancerNum}{integer: minimum number of Cancer types. Pan-Cancer, <=32}

\item{programNum}{integer: minimum number of target-predicting programs. <=7}

\item{program}{string: target-predicting programs. e.g., PITA, RNA22, miRmap, DIANA-microT, miRanda, PicTar and TargetScan}

\item{target}{gene name. e.g., TP53 "all" for downloading all regulatory data.}

\item{cellType}{cell type. e.g., HeLa. "all" for downloading all regulatory data.}
}
\value{
A data frame object

miRNAid	MiRBase microRNA ID, e.g., MIMAT0002175

miRNAname	Name of microRNA, e.g., hsa-miR-485-5p

geneID	ENSEMBL gene ID, e.g., ENSG00000141510

geneName	Name of gene, e.g., TP53

geneType	Type of gene, e.g., protein_coding

chromosome	The name of the chromosome, e.g., chr3, chrY

narrowStart	The starting position of the overlapped part of stacked target region (across target-predicting programs)

narrowEnd	The ending position of the overlapped part of stacked target region (across target-predicting programs)

broadStart	The starting position of the full range of stacked target region (across target-predicting programs)

broadEnd	The ending position of the full range of stacked target region (across target-predicting programs)

strand	Defines the strand. Either "+" or "-"

clipExpNum	Number of supporting CLIP-seq experiments

degraExpNum	Number of supporting degradome-seq experiments

RBP	Name of RNA binding protein

PITA	Number of target sites predicted by PITA

RNA22	Number of target sites predicted by RNA22

miRmap	Number of target sites predicted by miRmap

microT	Number of target sites predicted by microT

miRanda	Number of target sites predicted by miRanda

PicTar	Number of target sites predicted by PicTar

TargetScan	Number of target sites predicted by TargetScan

pancancerNum	Number of Cancer types (r<0, p-value<0.05)
}
\description{
Retrieve data of the miRNA-target interactions by intersecting the predicting target sites of miRNAs with binding sites of Ago protein
}
\examples{

tp53_vdr <-  mirna_target(target = c("TP53","VDR"), geneType = "mRNA")
h19 <- mirna_target(target = "H19", geneType = "lncRNA")
}
