% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_disease.R
\name{rbp_disease}
\alias{rbp_disease}
\title{Get Data for RBP and Somatic Gene Mutations}
\usage{
rbp_disease(assembly = "hg19", RBP = "all", tissue, disease, target)
}
\arguments{
\item{assembly}{=[genome version]: hg19}

\item{RBP}{=[protein name]. e.g., CBX7 ("all" for downloading all regulatory data)}

\item{tissue}{=[tissue name or organ name]. e.g., adrenal gland, breast}

\item{disease}{=[cancers and rare diseases in human]. e.g., Rosai-Dorfman disease, Crohn disease, Proteus syndrome}

\item{target}{=[gene name]. e.g., MYC ("all" for downloading all regulatory data)}
}
\value{
RBP	Name of RNA binding protein, e.g., ACIN1

geneID	ENSEMBL gene ID, e.g., ENSG00000136997

geneName	Name of gene, e.g., MYC

tissue	Tissue name or organ name, e.g., adrenal gland, breast

diseaseNum	The number of diseases

diseases	Name of cancers or rare diseases in human

diseaseCosmicID	The comma-separated mutations (COSMIC ID) corresponded to diseases

cosmicNum	The number of mutations (COSMIC ID)

sampleNum	Number of samples

mutTypeNum	Number of mutation types

clipExpNum	Number of supporting CLIP-seq experiments

clipNum	Number of CLIP-seq sites
}
\description{
@description Retrieve data for RBP-gene interactions and somatic mutations in human diseases
}
\examples{

disease_myc <- rbp_disease(tissue = c("prostate","breast"),disease = "carcinoma", target = "MYC")

}
