% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_motify.R
\name{rbp_motify}
\alias{rbp_motify}
\title{Filter the binding motifs of RBPs contained specific sequence pattern}
\usage{
rbp_motify(assembly = "hg19", motif, rankLimit = 10)
}
\arguments{
\item{assembly}{=[unique genome version]. hg19, mm10, dm6, ce10, sacCer3}

\item{motif}{=[motif pattern]. e.g., RRAC}

\item{rankLimit}{=[integer]. Rank limit of RNA binding motifs (default is 10)}
}
\value{
RBP Name of RNA binding protein, e.g., ACIN1

DatasetID

MotifRank

IdentifiedMotif

QueryMotif

targetPeakNum

TargetPercentage

p.value

p.value.ln

MotifMatrix

Region

CellLine.Tissue

Properties PAR-CLIP, Differentiated Glioma stem cells, rep1

MainAccession

SubAccession

Citation
}
\description{
Retrieve binding motifs of RBPs contained specific sequence pattern, which were de novo identified from their CLIP-seq peak data by Homer.
}
\examples{

mot_rrac <- rbp_motify(motif = "RRAC")

}
