% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_target.R
\name{rbp_target}
\alias{rbp_target}
\title{Get RBP Target Data}
\usage{
rbp_target(
  assembly = "hg19",
  geneType,
  RBP = "all",
  pancancerNum = 0,
  clipExpNum = 5,
  target,
  cellType = "all"
)
}
\arguments{
\item{assembly}{=[unique genome version]. hg19, mm10, dm6, ce10, sacCer3}

\item{geneType}{=[main gene type]: mRNA, lncRNA, pseudogene, circRNA, sncRNA}

\item{RBP}{=[protein name]: CBX7 ("all" for downloading all regulatory data)}

\item{pancancerNum}{=[integer]: minimum number of Cancer types (Pan-Cancer, <=32)}

\item{clipExpNum}{=[integer]: minimum number of supporting CLIP-seq experiments}

\item{target}{=[gene name]. e.g., TP53 ("all" for downloading all regulatory data)}

\item{cellType}{=[cell type]. e.g., HeLa ("all" for downloading all regulatory data)}
}
\value{
RBP	Name of RNA binding protein

geneID	ENSEMBL gene ID, e.g., ENSG00000148584

geneName	Name of gene, e.g., A1CF

geneType	Type of gene, e.g., protein_coding

clusterNum	The number of stacked peak regions

totalClipExpNum	The number of total CLIP-seq experiments

totalClipSiteNum	The number of total supporting binding sites from CLIP-seq experiments

clusterID	The ID of stacked peak regions

chromosome	The name of the chromosome, e.g., chr3, chrY

narrowStart	The starting position of the overlapped part of stacked target region (across target-predicting programs)

narrowEnd	The ending position of the overlapped part of stacked target region (across target-predicting programs)

broadStart	The starting position of the full range of stacked target region (across target-predicting programs)

broadEnd	The ending position of the full range of stacked target region (across target-predicting programs)

strand	Defines the strand. Either "+" or "-"

clipExpNum	Number of supporting CLIP-seq experiments

HepG2	The change (log2FC) of gene while knocking down RBP in HepG2 cell lines

K562	The change (log2FC) of gene while knocking down RBP in K562 cell lines

pancancerNum	Number of Cancer types (r<0, p-value<0.05)
}
\description{
Retrieve data of RBP-RNA interactions supported by the binding sites of RBPs derived from CLIP-seq data
}
\examples{
rbp_tp53 <- rbp_target(geneType = "mRNA", target = "TP53")
rbps <- rbp_target(geneType = "mRNA", target = c("TP53", "BRCA1"))
}
