% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-rna.R
\name{rna_rna}
\alias{rna_rna}
\title{Get the Data for NcRNA-RNA Interaction Network.}
\usage{
rna_rna(
  assembly = "hg19",
  geneType,
  RNA,
  interNum = 1,
  expNum = 1,
  cellType = "all"
)
}
\arguments{
\item{assembly}{=[genome version]: hg19, mm10}

\item{geneType}{=[main gene type]: mRNA, lncRNA, pseudogene, sncRNA, miRNA}

\item{RNA}{=[RNA name]. e.g., TP53 ("all" for downloading all regulatory data)}

\item{interNum}{=[integer]: miminum number of RNA-RNA interactions}

\item{expNum}{=[integer]: miminum number of experiments}

\item{cellType}{=[cell type]. e.g., HeLa ("all" for downloading all regulatory data)}
}
\value{
geneID	ENSEMBL gene ID, e.g., ENSG00000141510

geneName	Name of gene, e.g., TP53

geneType	Type of gene, e.g., protein_coding

pairGeneID	ENSEMBL gene ID, e.g., ENSG00000002834

pairGeneName	Name of gene, e.g., LASP1

pairGeneType	Type of gene, e.g., protein_coding

interactionNum	Number of RNA-RNA interactions

totalExpNum	Number of total experiments

totalSeqTypeNum	Number of total types of high-throughput sequencing

totalReadsNum	Number of total reads from all samples

interactionLocus	Pairing locus of RNA-RNA interactions

alignment	The consequential pairing of two genes

expNum	Number of supporting experiments

seqTypeNum	Number of types of high-throughput sequencing

readsNum	Number of reads

FreeEnergy	The minimum free energy of RNA-RNA pairs were calculated by RNAfold software

AlignScore(Smith-Waterman)	Calculated by Smith-Waterman algorithm. The analysis results with alignment score greater than 10 were kept and stored in ENCORI

pancancerNum	Number of Cancer types (r<0, p-value<0.05)
}
\description{
Retrieve the interaction network of ncRNA-RNA identified from high-throughput sequencing data of RNA-RNA interactome
}
\examples{
int_tp_brca<- rna_rna(geneType = "mRNA", RNA = c("TP53", "BRCA1"))
int_h19 <- rna_rna(geneType = "lncRNA", RNA = "H19")

}
